Extract Logging module
